Add union example
