+++ package.json
@@ -59,7 +59,7 @@
         "neo4j-driver": "^5.26.0",
         "prettier": "^3.4.1",
         "ts-jest": "^29.2.5",
-        "typedoc": "^0.27.0",
+        "typedoc": "^0.27.5",
         "typescript": "^5.6.3"
     }
 }
